+++ run_celery.sh
@@ -1,1 +1,2 @@
+#!/usr/bin/env bash
 celery -A config worker -l info