+++ run_celerybeat.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+celery -A config beat